python_backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
numpy==1.26.2
tensorflow==2.18.0
keras==3.10.0
Pillow==10.1.0
requests==2.31.0